+++ exporters/otlp/otlpmetric/otlpmetrichttp/internal/oconf/options_test.go
@@ -26,7 +26,6 @@ import (
 
 	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp/internal/envconfig"
 	"go.opentelemetry.io/otel/sdk/metric"
-	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 )
 
@@ -417,7 +416,7 @@ func TestConfigs(t *testing.T) {
 				// all" was set.
 				var undefinedKind metric.InstrumentKind
 				got := c.Metrics.AggregationSelector
-				assert.Equal(t, aggregation.Drop{}, got(undefinedKind))
+				assert.Equal(t, metric.AggregationDrop{}, got(undefinedKind))
 			},
 		},
 	}
@@ -443,8 +442,8 @@ func TestConfigs(t *testing.T) {
 	}
 }
 
-func dropSelector(metric.InstrumentKind) aggregation.Aggregation {
-	return aggregation.Drop{}
+func dropSelector(metric.InstrumentKind) metric.Aggregation {
+	return metric.AggregationDrop{}
 }
 
 func deltaSelector(metric.InstrumentKind) metricdata.Temporality {