+++ sdk/metric/reader_test.go
@@ -316,7 +316,7 @@ func TestDefaultAggregationSelector(t *testing.T) {
 	}
 
 	for _, ik := range iKinds {
-		assert.NoError(t, DefaultAggregationSelector(ik).Err(), ik)
+		assert.NoError(t, DefaultAggregationSelector(ik).err(), ik)
 	}
 }